--- src/models/GNNs.py
@@ -1,74 +1,0 @@
-
-import torch
-
-from models.MLP import MLP
-        
-class NRIMLP(torch.nn.Module):
-    def __init__(self, n_in, n_hid, n_out, do_prob=0., factor=True):
-        super(NRIMLP, self).__init__()
-
-        self.factor = factor
-
-        self.mlp1 = MLP(n_in, n_hid, n_hid, do_prob)
-        self.mlp2 = MLP(n_hid * 2, n_hid, n_hid, do_prob)
-        self.mlp3 = MLP(n_hid, n_hid, n_hid, do_prob)
-        if self.factor:
-            self.mlp4 = MLP(n_hid * 3, n_hid, n_hid, do_prob)
-            print("Using factor graph MLP encoder.")
-        else:
-            self.mlp4 = MLP(n_hid * 2, n_hid, n_hid, do_prob)
-            print("Using MLP encoder.")
-        self.fc_out = torch.nn.Linear(n_hid, n_out)
-        self.init_weights()
-
-    def init_weights(self):
-        for m in self.modules():
-            if isinstance(m, torch.nn.Linear):
-                torch.nn.init.xavier_normal(m.weight.data)
-                m.bias.data.fill_(0.1)
-
-    def edge2node(self, x, rel_rec):
-        # NOTE: Assumes that we have the same graph across all samples.
-        incoming = torch.matmul(rel_rec.t(), x)
-        return incoming / incoming.size(1)
-
-    def node2edge(self, x, rel_rec, rel_send):
-        # NOTE: Assumes that we have the same graph across all samples.
-        receivers = torch.matmul(rel_rec, x)
-        senders = torch.matmul(rel_send, x)
-        edges = torch.cat([senders, receivers], dim=2)
-        return edges
-
-    def forward(self, inputs, rel_rec, rel_send):
-        if len(inputs.shape) > 2:
-            # Input shape: [num_sims, num_atoms, num_timesteps, num_dims]
-            x = inputs.view(inputs.size(0), inputs.size(1), -1)
-            # New shape: [num_sims, num_atoms, num_timesteps*num_dims]
-        else:
-            x = inputs.view(1, inputs.shape[0], inputs.shape[1])
-        x = self.mlp1(x)  # 2-layer ELU net per node
-        x = self.node2edge(x, rel_rec, rel_send)
-        x = self.mlp2(x)
-        x_skip = x
-
-        if self.factor:
-            x = self.edge2node(x, rel_rec)
-            x = self.mlp3(x)
-            x = self.node2edge(x, rel_rec, rel_send)
-            x = torch.cat((x, x_skip), dim=2)  # Skip connection
-            x = self.mlp4(x)
-        else:
-            x = self.mlp3(x)
-            x = torch.cat((x, x_skip), dim=2)  # Skip connection
-            x = self.mlp4(x)
-
-        return self.fc_out(x)
-    
-class NRIMLPWrapper():
-    def __init__(self, n_in, n_hid=256, n_out=2):
-        self.model_name = "nrimlp"
-        self.n_epochs = 100
-
-        self.ModelClass = NRIMLP(n_in=n_in,
-                                 n_hid=n_hid,
-                                 n_out=n_out)
